Move callAppFunction stuff
